pathfinding pt. 3

+++ src/main/java/com/github/alexthe666/iceandfire/pathfinding/raycoms/pathjobs/ICustomSizeNavigator.java
@@ -3,6 +3,6 @@ package com.github.alexthe666.iceandfire.pathfinding.raycoms.pathjobs;
 public interface ICustomSizeNavigator {
 
     boolean isSmallerThanBlock();
-    int getXZNavSize();
+    float getXZNavSize();
     int getYNavSize();
 }
